done Place, MiniOffice and Orders for them + CRUD + Fixes + Added project for Data Layers

+++ CoCowork.DB/StoredProcedures/Room/Room_Insert.sql
@@ -1,7 +1,7 @@
 ﻿CREATE PROC dbo.Room_Insert
 	@Type int,
 	@AmountOfPeople int,
-	@PricePerHour int
+	@PricePerHour decimal(10,2)
 AS
 BEGIN
 	insert into dbo.Room
